test/web_capture/requirements.txt
playwright>=1.40.0
pydantic>=2.0.0
pytest>=7.0.0
pillow>=9.0.0
python-dotenv>=1.0.0
Jinja2>=3.0.0
rich>=13.0.0 